requirements.txt
fastapi
uvicorn
python-dotenv
watchfiles